Generate separate documentation for Promise-based library
We essentially have two different versions of the library – the one that
you get if you import `ably` (I’ll call this the “default” version), and
the one that you get if you import `ably/promises`. These libraries have
one important difference, which is that they have different exports:

- The default version has exports Rest and Realtime, which inherit from
  the Types.RestCallbacks and Types.RealtimeCallbacks classes.

- The Promise-based variant has exports Rest and Realtime, which inherit
  from the Types.RestPromise and Types.RealtimePromise classes.

So, to make sure that the exports described by our TypeDoc documentation
are consistent with the experience of importing `ably/promises`, we need
to generate separate documentation for the Promise-based variant.